# docs/api.yml
openapi: 3.0.3
info:
  title: Queue Service API
  description: >
    Прикладной интерфейс сервиса очереди на приобретение товара ограниченной
    доступности (кейс хакатона Авито «Авито Очередь»).


    Документ описывает исключительно конечные точки Queue Service. Внешняя
    система AvitoBackend — оформление заказа, проведение оплаты, хранение
    физического остатка — находится вне области кейса и здесь не описывается.
    Обоснование принятых решений приведено в `docs/design_context.md`, потоки
    сообщений — в `docs/diagrams.md`.


    Полноценная аутентификация вынесена за пределы кейса требованием `FR-D1`.
    Идентификатор пользователя передаётся заголовком `X-User-Id`. Для каналов
    WebSocket и Server-Sent Events, где браузерные конструкторы `WebSocket` и
    `EventSource` не позволяют задавать заголовки, предусмотрен параметр запроса
    `user_id`. Приоритет сохраняется за заголовком. В промышленной эксплуатации
    идентификатор извлекался бы из сессии, а не из запроса.


    Служебная конечная точка `POST /rights/{token}/events` вызывается внешней
    системой AvitoBackend и защищена разделяемым секретом в заголовке
    `X-Internal-Token`. Пустое значение секрета в конфигурации отключает
    проверку, что допустимо исключительно при локальной разработке.


    Интерфейс спроектирован в стиле REST: ресурсы именуются существительными
    (`members`, `rights`), действие определяется методом HTTP, а не глаголом в
    пути.


...
  version: "1.0.0"
servers:
...
security:
  - UserHeader: []
paths:
...
  /rights/{token}:
    get:
      tags: [rights]
      operationId: validateRight
      summary: Проверить действительность права
      description: >
        Вызывается браузером покупателя непосредственно перед обращением к
        AvitoBackend за созданием заказа. Код `200` возвращается исключительно в
        том случае, если право находится в состоянии `ACTIVE`, не истекло и
        принадлежит обратившемуся пользователю.


        Конечная точка реализует требования `FR-C1` и `FR-C2`. Ранее проверка
        предполагалась внутренним вызовом AvitoBackend → Queue Service на каждое
        создание заказа. Перенос проверки на сторону клиента и Queue Service
        снимает нагрузку с конечной точки оформления заказа: недействительные
        права — принадлежащие иному пользователю, истёкшие, использованные или
        несуществующие — отсекаются здесь.


        Чтение выполняется из кэша Redis; при промахе запись восстанавливается из
        PostgreSQL, являющегося источником истины.
      parameters:
        - $ref: '#/components/parameters/Token'
      responses:
        '200':
          description: Право действует и принадлежит обратившемуся пользователю
          content:
            application/json:
              schema:
                type: object
                required: [valid]
                properties:
                  valid:
                    type: boolean
                    enum: [true]
              example:
                valid: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Право принадлежит иному пользователю
        '404':
          description: >
            Право не существует либо находится в состоянии `EXPIRED` или `USED`.
            Три случая объединены намеренно: для вызывающей стороны они означают
            одно и то же, а их различение раскрывало бы сведения о существовании
            идентификатора.
...
components:
  securitySchemes:
    UserHeader:
      type: apiKey
      in: header
      name: X-User-Id
      description: >
        Идентификатор пользователя. Заменяет полноценную аутентификацию,
        вынесенную за пределы кейса требованием `FR-D1`.
    InternalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: >
        Разделяемый секрет межсервисного взаимодействия с AvitoBackend. Пустое
        значение в конфигурации отключает проверку.
  parameters:
...
    Token:
      name: token
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 9f3a1c7e-58d2-4a11-9f0b-2c6d4e7a1b83
...
  responses:
    Unauthorized:
      description: >
        Идентификатор пользователя не передан ни заголовком `X-User-Id`, ни
        параметром запроса `user_id`.
...